tests: split test_io.py
